Massively improved method
Well, after reading a bunch of blog posts and watch countless videos on recursion I stumbled across a post that reminded me that the 'slice' method existed. Now that I am not cutting the array in half through iteration, it's over 3x faster and more faithfully 'recursive.'  Let this be a reminder to future me:  learning doesn't happens in a vacuum, use all available resources!