Uploaded 'Journey' game
This is the first independent project I worked on when learning the basics of Ruby, in mid January. I haven't done any updating since, and it's quite a mess, but I figured I would upload it as a point of reference and to demonstrate the progress I've made. I think that little text based command-line games are a great way of familiarizing yourself with certain languages. For "Journey", I incorporated a few cool little features, like a 'dice rolling' game, and a 20 week journey at sea where you choose a character, and the event that occur are dependent on which character you chose.