Added save & load list functionality
Added the ability to save and load to-do lists, so you can actually close the program, re-open, and load a previously saved list... now, it's almost a useful thing. Have yet to figure out how to return an "error" message when the user types in an incorrect filename, instead of crashing the program. Small steps.